+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [FIXED] option determining whether json data written to the file (via DailyRotateFile transport) should be pretty printed or not, should be configurable
+
 ## 2.0.0
 
 * [ADDED] support for `Error.prototype.toLogger`